Byte code Conversion
As in interpreted languages, The Source code will be executed always whenever u run it, from the Interpreter level of execution. whereas in Compiler Languages, the source code is first compiled and converted into byte code, and then the compiler will execute the byte code whenever the source code is Executed. Whereas Here, Python is Multi-Paradigm Language. It Consists of both properties. To see it Manually, Execute the bytecode conversion steps as mentioned in the file.